refactor: Redis 캐시 반영 책임 전용 클래스로 LeafPointCacheUpdater 분리
